pyproject.toml
[build-system]
requires = ["setuptools>=64", "setuptools_scm[toml]>=8"]
build-backend = "setuptools.build_meta"

[project]
name = "pdftools"
classifiers = [
    "Development Status :: 3 - Alpha",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
description = "Importable tools for the PDF beamline at NSLS-II"
dependencies = [
    "bluesky",
    "ophyd-async==0.17.0a3",
    "tiled[client]"
]
dynamic = ["version"]
license.file = "LICENSE"
readme = "README.md"
requires-python = ">=3.11"

[dependency-groups]
dev = [
    "ophyd-async[ca, pva, sim]>=0.17a3",
    "inflection",
    "import-linter",
    "numpydoc",
    "pre-commit",
    "pyright",
    "pytest",
    "pytest-asyncio>=1.3.0,<2",
    "pytest-cov",
    "pytest-faulthandler",
    "pytest-forked",
    "pytest-insubprocess",
    "pytest-rerunfailures",
    "pytest-timeout>=2.4.0,<3",
    "ruff",
    "scanspec>=1.0a1",
    "types-mock",
    "types-pyyaml",
    "tiled[client,server]>=0.2.9",
    "bluesky-tiled-plugins>=2.0.2",
    "h5py>=3.16.0",
    "ipython",
]

[project.scripts]

[project.urls]
GitHub = "https://github.com/NSLS2/pdftools"

[[project.authors]] # Further authors may be added by duplicating this section
email = "jwlodek@bnl.gov"
name = "Jakub Wlodek"


[tool.setuptools_scm]
version_file = "src/pdftools/_version.py"

[tool.pyright]
typeCheckingMode = "standard"
reportMissingImports = false  # Ignore missing stubs in imported modules


[tool.pytest.ini_options]
# Run pytest with all our checkers, and don't spam us with massive tracebacks on error
addopts = """
    --tb=native -vv --strict-markers --doctest-modules
    --doctest-glob="*.rst" --doctest-glob="*.md"
    --ignore=docs/examples
    """
# https://iscinumpy.gitlab.io/post/bound-version-constraints/#watch-for-warnings
filterwarnings = ["error"]
# Doctest python code in docs, python code in src docstrings, test functions in tests
testpaths = "docs src tests"
log_format = "%(asctime)s,%(msecs)03d %(levelname)s (%(threadName)s) %(message)s"
log_date_format = "%H:%M:%S"

asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
timeout = 60

# tox must currently be configured via an embedded ini string
# See: https://github.com/tox-dev/tox/issues/999
[tool.tox]
legacy_tox_ini = """
[tox]
skipsdist=True

[testenv:{pre-commit,type-checking,tests}]
# Don't create a virtualenv for the command, requires tox-direct plugin
direct = True
passenv = *
allowlist_externals =
    pytest
    pre-commit
    pyright
commands =
    pre-commit: pre-commit run --all-files --show-diff-on-failure {posargs}
    type-checking: pyright src tests {posargs}
    tests: pytest --cov=pdftools --cov-report term --cov-report xml:cov.xml {posargs}
"""

[tool.ruff]
src = ["src", "tests"]
line-length = 88
lint.select = [
    "B",   # flake8-bugbear - https://docs.astral.sh/ruff/rules/#flake8-bugbear-b
    "C4",  # flake8-comprehensions - https://docs.astral.sh/ruff/rules/#flake8-comprehensions-c4
    "E",   # pycodestyle errors - https://docs.astral.sh/ruff/rules/#error-e
    "F",   # pyflakes rules - https://docs.astral.sh/ruff/rules/#pyflakes-f
    "W",   # pycodestyle warnings - https://docs.astral.sh/ruff/rules/#warning-w
    "I",   # isort - https://docs.astral.sh/ruff/rules/#isort-i
    "UP",  # pyupgrade - https://docs.astral.sh/ruff/rules/#pyupgrade-up
    "SLF", # self - https://docs.astral.sh/ruff/settings/#lintflake8-self
]
exclude = [
    "src/pdftools/_version.py",
]

[tool.ruff.lint.per-file-ignores]
# By default, private member access is allowed in tests
# See https://github.com/DiamondLightSource/python-copier-template/issues/154
# Remove this line to forbid private member access in tests
"tests/**/*" = ["SLF001"]

[tool.importlinter]
root_package = "pdftools"


[tool.pixi.workspace]
channels = ["conda-forge"]
platforms = ["linux-64"]

[tool.pixi.dependencies]
python = "*"

[tool.pixi.pypi-dependencies]
pdftools = { path = ".", editable = true }

[tool.pixi.feature.py311.dependencies]
python = "3.11.*"

[tool.pixi.feature.py312.dependencies]
python = "3.12.*"

[tool.pixi.feature.py313.dependencies]
python = "3.13.*"

[tool.pixi.feature.py314.dependencies]
python = "3.14.*"

[tool.pixi.environments]
default = { features = ["dev"], solve-group = "default" }
py311 = { features = ["dev", "py311"], solve-group = "py311" }
py312 = { features = ["dev", "py312"], solve-group = "py312" }
py313 = { features = ["dev", "py313"], solve-group = "py313" }
py314 = { features = ["dev", "py314"], solve-group = "py314" }

[tool.pixi.tasks]
tests = "pytest --cov=src/pdftools --cov-report=term --cov-report=xml:coverage.xml"
lint = "pre-commit run --all-files"
type-check = "pyright src tests"
